+++ src_new/aliases/levi_ataxia/for_levi/levi_062424/lists/002_Display_Mobs_To_Add.lua
@@ -21,7 +21,7 @@ end
 if ataxia.denizensHere and ataxiaBasher.targetList[gmcp.Room.Info.area] then
 	ataxiaEcho("Displaying mobs currently in room, but not in target list. Click to add.")
 	for i,v in pairs(ataxia.denizensHere) do
-		if not table.contains(ataxiaBasher.targetList[gmcp.Room.Info.area], v) and not table.contains(ataxiaBasher.mobIgnore, v) then
+		if not table.contains(ataxiaBasher.targetList[gmcp.Room.Info.area], v) and not table.contains(ataxiaBasher.mobIgnore, v) and not ataxiaBasher_isOwnDenizen(v) then
 			cecho("\n  <green>+ ")
 			fg("white")
 			echoLink(v, [[ataxiaBasher_addmob("]]..v..[[")]], "Add "..v.." to denizen list.", true)